% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_ambient_profile}
\alias{annotate_ambient_profile}
\title{Add average of ambient feature reads to rowData}
\usage{
annotate_ambient_profile(
  sce,
  ambient_indices,
  name = "ambient_profile",
  alt_exp = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Add average of ambient feature reads to rowData
}
\examples{
NULL
}
