% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cdr3.R
\name{annotate_cdr3}
\alias{annotate_cdr3}
\title{Annotate CDR3 sequences stored in colData(sce)$vdj based on a reference dataset (for example, Ag-specific data)}
\usage{
annotate_cdr3(
  sce,
  reference,
  reference_cdr3_column,
  reference_annotation_column,
  chain_match = "TRB",
  max_dist = 1,
  method = "hamming"
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{reference}{reference dataset for use in annotating query sequences}

\item{reference_cdr3_column}{column in reference which contains the CDR3 region}

\item{reference_annotation_column}{column in reference to annotate onto query dataset}

\item{chain_match}{chain(s) to match (for example, c("TRA", "TRB"))}

\item{max_dist}{maximum distance to allow between query and reference sequence}

\item{method}{distance metric- one of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"}
}
\value{
SingleCellExperiment object with additional column with the same name as reference_annotation_column
}
\description{
Annotate CDR3 sequences stored in colData(sce)$vdj based on a reference dataset (for example, Ag-specific data)
}
\examples{
NULL
}
