% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chain_count.R
\name{annotate_chain_count}
\alias{annotate_chain_count}
\title{Annotate number of TRA, TRB, IGL, and IGH reads per barcode}
\usage{
annotate_chain_count(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a column named vdj containing contig_annotations in colData}
}
\value{
SingleCellExperiment object with new count columns in colData
}
\description{
Columns count_TRA, count_TRB, count_IGL, count_IGH, and count_TRA_TRB_IGL_IGH (underscore delimited concatenation of previous 4 columns) are added to colData(sce)
}
\examples{
NULL
}
