% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clonotype_count.R
\name{annotate_clonotype_count}
\alias{annotate_clonotype_count}
\title{Annotate number of occurrences of a given clonotype in a sample}
\usage{
annotate_clonotype_count(sce, clonotype_name = "clonotype", group_by = c())
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{clonotype_name}{name of column in colData containing clonotype - can be generated using `assign_clonotypes`.}

\item{group_by}{columns to group repertoires by - only counts clonotypes within groups}
}
\value{
SingleCellExperiment object with added columns to colData including "clonotype", "clonotype_count", and "clonotype_count_2:10"
}
\description{
Annotate number of occurrences of a given clonotype in a sample
}
\examples{
NULL
}
