% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_num_cells_expr}
\alias{annotate_num_cells_expr}
\title{Add number of cells with nonzero expression for each gene to rowData(sce)$num_cells_expr}
\usage{
annotate_num_cells_expr(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Add number of cells with nonzero expression for each gene to rowData(sce)$num_cells_expr
}
\examples{
NULL
}
