% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_pct_gene_set}
\alias{annotate_pct_gene_set}
\title{Add percent of gene set defined by a regular expression pattern to colData(sce)}
\usage{
annotate_pct_gene_set(sce, pattern, name, alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{pattern}{Regular expression for gene names to sum up expression}

\item{name}{Column name to save data in colData(sce)}
}
\value{

}
\description{
Add percent of gene set defined by a regular expression pattern to colData(sce)
}
\examples{
NULL
}
