% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_pct_total_reads}
\alias{annotate_pct_total_reads}
\title{Add percentage of total reads that each gene takes up across the whole dataset to rowData(sce)$pct_reads}
\usage{
annotate_pct_total_reads(sce, name = "pct_reads", alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Add percentage of total reads that each gene takes up across the whole dataset to rowData(sce)$pct_reads
}
\examples{
NULL
}
