% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_total_umi_count}
\alias{annotate_total_umi_count}
\title{Add number of total umis per barcode to colData(sce)$total_umi}
\usage{
annotate_total_umi_count(sce, name = "total_umi", alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Add number of total umis per barcode to colData(sce)$total_umi
}
\examples{
NULL
}
