% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_clonotypes.R
\name{assign_clonotypes}
\alias{assign_clonotypes}
\title{Function to use for assigning clonotypes based on custom definition}
\usage{
assign_clonotypes(
  sce,
  clonotype_fields = c("v_gene", "cdr3", "j_gene"),
  chains = c("TRA", "TRB"),
  tra_range = c(0, 1),
  trb_range = c(0, 1),
  igl_range = c(0, 2),
  igh_range = c(0, 1)
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{clonotype_fields}{character vector indicating fields to include in clonotype definition (e.g., c("v_gene", "cdr3", "j_gene", etc.))}

\item{chains}{character vector indicating chains which to include in clonotype definition(e.g., c("TRA", "TRB"))}

\item{tra/trb/igl/igh_range}{integer vector of length 2 with minimum and maximum number of unique chains (inclusive)}
}
\value{
SingleCellExperiment object with added columns to colData including "clonotype", "clonotype_count", and "clonotype_count_2:10"
}
\description{
Function to use for assigning clonotypes based on custom definition
}
\examples{
NULL
}
