% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Cache and retrieve intermediate steps if path exists}
\usage{
cache(file_path, fn)
}
\arguments{
\item{file_path}{file path at which to cache}

\item{fn}{function whose result to cache}
}
\value{
cached object
}
\description{
Cache and retrieve intermediate steps if path exists
}
\examples{
NULL
}
