% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_evenness_profile.R
\name{compute_evenness_profile_long}
\alias{compute_evenness_profile_long}
\title{Compute tidied data frame of evenness profiles per group based on clonotype frequency distributions}
\usage{
compute_evenness_profile_long(
  sce_list,
  group_by = c(),
  min_alpha = 0,
  max_alpha = 10,
  alpha_steps = 51
)
}
\arguments{
\item{sce_list}{Named list of SingleCellExperiment objects}

\item{group_by}{Character vector of column names from colData to use to construct groups for clonotype counting}

\item{min_alpha}{Minimum Renyi entropy alpha}

\item{max_alpha}{Maximum Renyi entropy alpha}

\item{alpha_steps}{Number of steps in between min_alpha and max_alpha}
}
\value{

}
\description{
Compute tidied data frame of evenness profiles per group based on clonotype frequency distributions
}
\examples{
NULL
}
