% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_identity_frequency.R
\name{convert_identity_frequency_matrix_to_long}
\alias{convert_identity_frequency_matrix_to_long}
\title{Convert an identity frequency matrix}
\usage{
convert_identity_frequency_matrix_to_long(matrix)
}
\arguments{
\item{matrix}{result from encode_*_identity_frequency_long}
}
\value{

}
\description{
Convert an identity frequency matrix
}
\examples{
NULL
}
