% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_identity_frequency.R
\name{.encode_identity_frequency_matrix}
\alias{.encode_identity_frequency_matrix}
\title{Helper function to encode identity frequency matrix given data frame}
\usage{
.encode_identity_frequency_matrix(data, attributes, group_by, normalize)
}
\arguments{
\item{data}{data frame to encode}

\item{attributes}{columns to use to define features}

\item{group_by}{columns to use to define groups}

\item{normalize}{normalization method}
}
\value{

}
\description{
Helper function to encode identity frequency matrix given data frame
}
\examples{
NULL
}
