% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_barcode.R
\name{.filter_ambient_barcode_empty_drops}
\alias{.filter_ambient_barcode_empty_drops}
\title{Get filter for cells based on emptyDrops algorithm}
\usage{
.filter_ambient_barcode_empty_drops(sce, fdr, lower_umi_limit, n_iters)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{fdr}{upper limit for FDR}

\item{lower_limit_umi}{lower limit for UMI, below which all barcodes are assumed to be ambient}
}
\value{
Boolean filter with TRUE for cells and FALSE for ambient RNA. Attribute named data containing output of emptyDrops function
}
\description{
Get filter for cells based on emptyDrops algorithm
}
\examples{
NULL
}
\keyword{internal}
