% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_barcode.R
\name{.filter_ambient_barcode_inflection}
\alias{.filter_ambient_barcode_inflection}
\title{Get filter for cells based on inflection point}
\usage{
.filter_ambient_barcode_inflection(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
Boolean filter with TRUE for cells and FALSE for ambient RNA. Attribute named value containing total UMI at inflection point.
}
\description{
Get filter for cells based on inflection point
}
\examples{
NULL
}
