% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_barcode.R
\name{.filter_ambient_barcode_knee}
\alias{.filter_ambient_barcode_knee}
\title{Get filter for cells based on knee point}
\usage{
.filter_ambient_barcode_knee(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
Boolean filter with TRUE for cells and FALSE for ambient RNA. Attribute named value containing total UMI at knee point.
}
\description{
Get filter for cells based on knee point
}
\examples{
NULL
}
\keyword{internal}
