% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{.get_cell_annotations}
\alias{.get_cell_annotations}
\title{Get cell annotations for feature_heatmap}
\usage{
.get_cell_annotations(sce_list, cell_features)
}
\arguments{
\item{cell_features}{}
}
\value{

}
\description{
Get cell annotations for feature_heatmap
}
