% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_features.R
\name{.get_cell_features_assay_explicit_exp}
\alias{.get_cell_features_assay_explicit_exp}
\title{Get feature from assay (from alternate experiment)}
\usage{
.get_cell_features_assay_explicit_exp(sce, features, assay)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{features}{Row names from assay object of relevant alternate experiment}

\item{assay}{Assay to use (counts, logcounts, etc.)}

\item{alt_exp}{Name of the altExp to use (if any)}
}
\value{
data.frame with rows as cells and columns as features
}
\description{
Get feature from assay (from alternate experiment)
}
\examples{
NULL
}
\keyword{internal}
