% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_features.R
\name{.get_cell_features_coldata}
\alias{.get_cell_features_coldata}
\title{Get feature from column metadata}
\usage{
.get_cell_features_coldata(sce, features)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{features}{Column names from column metadata}
}
\value{
data.frame with rows as cells and columns as features
}
\description{
Get feature from column metadata
}
\examples{
NULL
}
