% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{.get_feature_annotations}
\alias{.get_feature_annotations}
\title{Get feature annotations for feature_heatmap}
\usage{
.get_feature_annotations(sce_list, features, alt_exp)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{features}{columns from rowData}

\item{alt_exp}{altExp to plot}
}
\value{

}
\description{
Get feature annotations for feature_heatmap
}
\examples{
NULL
}
