% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduced_dimensions.R
\name{.get_reduced_dims}
\alias{.get_reduced_dims}
\title{Get reduced dimensions of object}
\usage{
.get_reduced_dims(sce, type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{type}{Name of reduction type in reducedDims}
}
\value{

}
\description{
Get reduced dimensions of object
}
\examples{
NULL
}
