% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_qc.R
\name{.prepare_barcode_counts}
\alias{.prepare_barcode_counts}
\title{Prepare count annotations for number of cells passing each filtering method}
\usage{
.prepare_barcode_counts(sce, sample_name, ambient_rna_filter, facets)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{sample_name}{name of sample}

\item{ambient_rna_filter}{result from filter_ambient_rna function}

\item{facets}{columns to group by in plot}
}
\value{

}
\description{
Prepare count annotations for number of cells passing each filtering method
}
\examples{
NULL
}
\keyword{internal}
