% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_qc.R
\name{.prepare_barcode_cutoffs}
\alias{.prepare_barcode_cutoffs}
\title{Prepare cutoff values for each method}
\usage{
.prepare_barcode_cutoffs(sce, sample_name, ambient_rna_filter)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{sample_name}{Name of sample}

\item{ambient_rna_filter}{Result from filter_ambient_rna function}
}
\value{

}
\description{
Prepare cutoff values for each method
}
\examples{
NULL
}
\keyword{internal}
