% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_qc.R
\name{.prepare_barcode_plot_data}
\alias{.prepare_barcode_plot_data}
\title{Prepare data for plotting barcode QC plot}
\usage{
.prepare_barcode_plot_data(sce, sample_name, ambient_rna_filter, metadata)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{sample_name}{name of sample}

\item{ambient_rna_filter}{result from filter_ambient_rna function}
}
\value{

}
\description{
Prepare data for plotting barcode QC plot
}
\examples{
NULL
}
\keyword{internal}
