% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gex_qc.R
\name{.prepare_gex_bivariate_counts}
\alias{.prepare_gex_bivariate_counts}
\title{Prepare counts of cells that meet filter criteria to annotate based on two filters}
\usage{
.prepare_gex_bivariate_counts(
  data,
  x,
  y,
  x_filter,
  y_filter,
  x_log,
  y_log,
  sample_name,
  facets
)
}
\arguments{
\item{data}{Result from .prepare_gex_data}

\item{x}{Feature from colData to plot distribution of}

\item{y}{Feature from colData to plot distribution of}

\item{x_filter}{Filter for a sample from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{y_filter}{Filter for a sample from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{x_log}{Boolean to plot x on log-scale or not}

\item{y_log}{Boolean to plot y on log-scale or not}

\item{sample_name}{Sample name}

\item{facets}{Features from colData to facet on}
}
\value{

}
\description{
Prepare counts of cells that meet filter criteria to annotate based on two filters
}
\examples{
NULL
}
\keyword{internal}
