% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gex_qc.R
\name{.prepare_gex_data}
\alias{.prepare_gex_data}
\title{Prepare data frame to plot for gene expression related QC plots}
\usage{
.prepare_gex_data(sce, sample_name)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{sample_name}{Name of sample}
}
\value{

}
\description{
Prepare data frame to plot for gene expression related QC plots
}
\examples{
NULL
}
\keyword{internal}
