% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{.prepare_volcano_annotations}
\alias{.prepare_volcano_annotations}
\title{Prepare annotations for volcano plot}
\usage{
.prepare_volcano_annotations(
  data,
  fold_change,
  p_value,
  label,
  n_annotate_top,
  annotations,
  annotations_if_threshold,
  p_value_threshold,
  fold_change_threshold,
  p_value_rank_threshold,
  facet_rows,
  facet_columns
)
}
\arguments{
\item{data}{differential expression results}

\item{fold_change}{column containing fold change data (x-axis)}

\item{p_value}{column containing p-value data (y-axis is -log10(p_value))}

\item{label}{column for text annotation}

\item{n_annotate_top}{number of top DE genes to annotate}

\item{annotations}{features to be annotated}

\item{annotations_if_threshold}{features to be annotated iff they meet the thresholds}

\item{p_value_threshold}{maximum p-value}

\item{fold_change_threshold}{minimum absolute fold change threshold}

\item{p_value_rank_threshold}{maximum p-value rank threshold}

\item{facet_rows}{columns for faceting by row}

\item{facet_columns}{columns for faceting by column}
}
\value{

}
\description{
Prepare annotations for volcano plot
}
\examples{
NULL
}
\keyword{internal}
