% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{.prepare_volcano_top_annotations}
\alias{.prepare_volcano_top_annotations}
\title{Prepare the top N annotations for volcano plot}
\usage{
.prepare_volcano_top_annotations(
  data,
  n_annotate_top,
  fold_change,
  p_value,
  facet_rows,
  facet_columns
)
}
\arguments{
\item{data}{differential expression results}

\item{n_annotate_top}{number of top DE genes to annotate}

\item{fold_change}{column containing fold change data (x-axis)}

\item{p_value}{column containing p-value data (y-axis is -log10(p_value))}

\item{facet_rows}{columns for faceting by row}

\item{facet_columns}{columns for faceting by column}
}
\value{

}
\description{
Prepare the top N annotations for volcano plot
}
\examples{
NULL
}
