% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{.read_10x_gene_expr}
\alias{.read_10x_gene_expr}
\title{Read gene expression matrix from 10X data}
\usage{
.read_10x_gene_expr(path)
}
\arguments{
\item{path}{path to gene expression data - this path must contain mtx, features, and barcodes file}
}
\value{
SingleCellExperiment object
}
\description{
Read gene expression matrix from 10X data
}
\examples{
NULL
}
\keyword{internal}
