% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{.read_10x_vdj}
\alias{.read_10x_vdj}
\title{Read VDJ information from 10X data}
\usage{
.read_10x_vdj(path)
}
\arguments{
\item{path}{path to vdj data - this path must contain all_contig_annotations.csv}
}
\value{
DataFrame with vdj information
}
\description{
Read VDJ information from 10X data
}
\examples{
NULL
}
\keyword{internal}
