% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{.sce_to_assay}
\alias{.sce_to_assay}
\title{Convert SingleCellExperiment object to Seurat assay}
\usage{
.sce_to_assay(sce, return_assay = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Convert SingleCellExperiment object to Seurat assay
}
\examples{
NULL
}
