% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{.seurat_assay_to_sce}
\alias{.seurat_assay_to_sce}
\title{Convert Seurat assay to SingleCellExperiment}
\usage{
.seurat_assay_to_sce(seurat, assay)
}
\arguments{
\item{seurat}{Seurat object}

\item{assay}{assay to convert}
}
\value{

}
\description{
Convert Seurat assay to SingleCellExperiment
}
\examples{
NULL
}
\keyword{internal}
