% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_identity_frequency.R
\name{encode_vdj_identity_frequency_long}
\alias{encode_vdj_identity_frequency_long}
\title{Encode frequencies of (combinations) of values in columns from colData into long data frame}
\usage{
encode_vdj_identity_frequency_long(
  sce_list,
  attributes,
  group_by = c(),
  normalize = "none"
)
}
\arguments{
\item{sce_list}{List of SCE objects, all containing relevant attributes in colData(sce)$vdj as columns}

\item{attributes}{Column names from unnested colData(sce)$vdj which to encode}

\item{group_by}{Column names from colData which to group by (compute frequencies per group)}

\item{normalize}{Normalization method - either "none" or "relative_frequency"}
}
\value{

}
\description{
Encode frequencies of (combinations) of values in columns from colData into long data frame
}
\examples{
NULL
}
