% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_barcode.R
\name{filter_ambient_barcode}
\alias{filter_ambient_barcode}
\title{Get filter to remove ambient RNA barcodes based on the combination of multiple methods:}
\usage{
filter_ambient_barcode(sce, n_cells, fdr = 0.01, lower_umi_limit = 100)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{n_cells}{Number of input cells (used in cellranger cutoff computation)}

\item{fdr}{FDR threshold for keeping a cell for emptyDrops}

\item{lower_umi_limit}{Number of total UMIs below which a barcode is assumed to contain ambient RNA, used in emptyDrops method only}
}
\value{
List for each method in containing filter for method and attributes with details specific to each method.
}
\description{
inflection - all barcodes below the inflection point in the UMI rank vs total UMI plot are called ambient
knee - all barcodes below the knee in the UMI rank vs total UMI plot are called ambient
cellranger (v2) - all barcodes below 10% of the 99th percentile of total UMIs in the top n_cells barcodes are called ambient
empty_drops - uses the algorithm from DropletUtils::emptyDrops to compare each barcode to the profile of barcodes below
}
\examples{
NULL
}
