% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_barcode.R
\name{filter_ambient_barcode_cellranger}
\alias{filter_ambient_barcode_cellranger}
\title{Get filter for cells based on inflection point}
\usage{
filter_ambient_barcode_cellranger(sce, n_cells)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{n_cells}{Number of cells expected}
}
\value{
Boolean filter with TRUE for cells and FALSE for ambient RNA. Attribute named value containing total UMI at the cutoff point.
}
\description{
Get filter for cells based on inflection point
}
\examples{
NULL
}
