% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gex_outliers.R
\name{filter_total_umi}
\alias{filter_total_umi}
\title{Get filter for total umis per barcode based on log transformed values outside (either above, below, or both based on type parameter) nmads median absolute deviations from the median}
\usage{
filter_total_umi(sce, nmads, type, column = "total_umi")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{nmads}{number of median absolute deviations from the median for cutoff}

\item{type}{either "both", "lower", or "higher", referring to which side to filter on}

\item{column}{column in colData containing total UMI count}
}
\value{
Boolean filter with TRUE for cells and FALSE for outliers. Contains attribute with number of median absolute deviations specified.
}
\description{
Wrapper around scater::isOutlier, arguments type and nmads refer to the same thing. For total_umi, uses log transformed values for deciding cutoffs.
}
\examples{
NULL
}
