% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vdj_chain_count.R
\name{filter_vdj_chain_count}
\alias{filter_vdj_chain_count}
\title{Get filter for barcodes based on the number of chains for TRA, TRB, IGL, IGK, and IGH present}
\usage{
filter_vdj_chain_count(
  sce,
  tra_range = c(0, 2),
  trb_range = c(0, 1),
  igl_range = c(0, 2),
  igk_range = c(0, 2),
  igh_range = c(0, 1),
  tcr_and_bcr_allowed = TRUE,
  only_filter_if_any_vdj_hits = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{tcr_and_bcr_allowed}{whether reads to both TRA/TRB locus and IGL/IGK/IGH locus are allowed on a single barcode - if FALSE, barcodes with reads to both will be removed}

\item{only_filter_if_any_vdj_hits}{whether barcodes should only be filtered if there are any TRA/TRB/IGL/IGK/IGH reads at all - for diverse populations with non B/T cells, should be set to TRUE in order to avoid meaningless filtering on non B/T cells}

\item{min_tra}{range of unique TRA sequences allowed per barcode. Barcodes with less than the first number and greater than the second number will be removed}

\item{min_trb}{range of unique TRB sequences allowed per barcode. Barcodes with less than the first number and greater than the second number will be removed}

\item{min_igl}{range of unique IGL sequences allowed per barcode. Barcodes with less than the first number and greater than the second number will be removed}

\item{min_igh}{range of unique IGH sequences allowed per barcode. Barcodes with less than the first number and greater than the second number will be removed}
}
\value{
Boolean filter with TRUE for cells kept and FALSE for cells outside the allowed ranges.
}
\description{
Wrapper around scater::isOutlier, arguments type and nmads refer to the same thing. For mitochondrial reads, uses non-log transformed values for deciding cutoffs.
}
\examples{
NULL
}
