% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preprocessing_report.R
\name{generate_preprocessing_report}
\alias{generate_preprocessing_report}
\title{Generate standardized preprocessing report with filtering based on removal of ambient RNA barcodes, gene expression outliers, and (some) doublets based on VDJ}
\usage{
generate_preprocessing_report(
  sce,
  sample_name = "sample",
  n_cells = 5000,
  fdr = 0.01,
  lower_umi_limit = 100,
  is_cell_criteria = c("inflection", "empty_drops"),
  is_cell_boolean_operator = "&",
  total_umi_filter_nmads = 2,
  total_umi_filter_type = "both",
  n_genes_expr_filter_nmads = 2,
  n_genes_expr_filter_type = "both",
  pct_mito_filter_nmads = 2,
  pct_mito_filter_type = "higher",
  tra_range = c(0, 2),
  trb_range = c(0, 1),
  igl_range = c(0, 2),
  igh_range = c(0, 1),
  tcr_and_bcr_allowed = FALSE,
  only_filter_if_any_vdj_hits = TRUE,
  result_path = getwd()
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{sample_name}{Name of sample for saving object}

\item{n_cells}{Number of input cells (used in cellranger cutoff computation)}

\item{fdr}{FDR threshold for keeping a cell for emptyDrops}

\item{lower_umi_limit}{Number of total UMIs below which a barcode is assumed to contain ambient RNA, used in emptyDrops method only}

\item{is_cell_criteria}{inflection, empty_drops, knee, and/or cellranger, cells are used based on combination of these}

\item{is_cell_boolean_operator}{Boolean operator to combine the criteria - either & or |}
}
\value{
Combined boolean filter with TRUE for cells and FALSE for ambient RNA. Attributes for each method in criteria containing details specific to each method.
}
\description{
Generate standardized preprocessing report with filtering based on removal of ambient RNA barcodes, gene expression outliers, and (some) doublets based on VDJ
}
\examples{
NULL
}
