% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_features.R
\name{get_cell_features}
\alias{get_cell_features}
\title{Get feature from assay data, colData, or reducedDims at once from main experiment or alternate experiments}
\usage{
get_cell_features(sce, features, assay, alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{features}{Names of features desired. It will first check row names of the assay data, then column names of the colData, then column names of each of the reducedDims slots. Only those found will be returned}

\item{assay}{Assay to use (counts, logcounts, etc.)}

\item{alt_exp}{Name of the altExp to use (if any)}
}
\value{
data.frame with rows as cells and columns as features
}
\description{
Get feature from assay data, colData, or reducedDims at once from main experiment or alternate experiments
}
\examples{
NULL
}
