% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_features.R
\name{get_cell_features_reduced_dimensions}
\alias{get_cell_features_reduced_dimensions}
\title{Get feature from reduced dimensional representation}
\usage{
get_cell_features_reduced_dimensions(sce, features)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{features}{Names of dimensions to retreive in format of [red_dim_name]_[dimension_number] (e.g. PCA_1)}
}
\value{
data.frame with rows as cells and columns as features
}
\description{
Get feature from reduced dimensional representation
}
\examples{
NULL
}
