% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_multi_sample_pbmc_10k}
\alias{get_multi_sample_pbmc_10k}
\title{Get modified version of PBMC 10k data from 10X}
\usage{
get_multi_sample_pbmc_10k()
}
\value{

}
\description{
Get modified version of PBMC 10k data from 10X
}
\examples{
data = get_pbmc_10k_data()
}
