% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbmc_10k_data.R
\name{get_pbmc_10k_data}
\alias{get_pbmc_10k_data}
\title{Get modified version of PBMC 10K data from 10X}
\usage{
get_pbmc_10k_data()
}
\value{

}
\description{
Get modified version of PBMC 10K data from 10X
}
