% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_pbmc_5k_nextgem}
\alias{get_pbmc_5k_nextgem}
\title{Get PBMC 5k data using v3 chemistry}
\usage{
get_pbmc_5k_nextgem()
}
\value{

}
\description{
Get PBMC 5k data using v3 chemistry
}
\examples{
data = get_pbmc_5k_nextgem()
}
