% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assay_data.R
\name{get_row_data}
\alias{get_row_data}
\title{Get rowData from either the main experiment or altExps}
\usage{
get_row_data(sce, assay, alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{alt_exp}{alternate experiment to get rowData from}
}
\value{

}
\description{
Get rowData from either the main experiment or altExps
}
\examples{
NULL
}
