% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_qc.R
\name{plot_barcode_qc}
\alias{plot_barcode_qc}
\title{QC plot of UMI rank vs total number of UMIs per barcode}
\usage{
plot_barcode_qc(
  sce_list,
  ambient_rna_filters,
  color = "empty_drops",
  shape = NULL,
  facet_rows = NULL,
  facet_columns = NULL,
  facet_type = "grid",
  point_size = 1,
  text_size = 2,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{ambient_rna_filters}{list of results from filter_ambient_rna function, one for each SCE object in sce_list - must be the same names and order}

\item{color}{column from colData that is in all objects in sce_list}

\item{shape}{column from colData that is in all objects in sce_list}

\item{facet_rows}{columns to facet on}

\item{facet_columns}{columns to facet on}

\item{facet_type}{either "wrap" or "grid", same as ggplot}

\item{text_size}{font size for annotations}

\item{...}{other params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{
ggplot with plot of UMI rank vs total number of UMIs per barcode
}
\description{
QC plot of UMI rank vs total number of UMIs per barcode
}
\examples{
NULL
}
