% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot feature distributions from SingleCellExperiment objects}
\usage{
plot_distributions(
  sce_list,
  assay,
  alt_exp = NULL,
  features,
  x,
  color = x,
  facet_rows = c(".feature"),
  facet_columns = c(".sample"),
  ...
)
}
\arguments{
\item{sce_list}{List of SingleCellExperiment objects}

\item{assay}{Assay to obtain data from (counts, logcounts, etc)}

\item{alt_exp}{Alternate experiment to obtain data from}

\item{features}{Numeric features to plot - can be from reducedDims, colData, or assay data}

\item{x}{colData variable to plot on x-axis}

\item{color}{colData variable to color points by}

\item{facet_rows}{Variables from colData to facet on, can also include ".sample" or ".feature" as described below}

\item{facet_columns}{Variables from colData to facet on, can also include ".sample" or ".feature" as described below}

\item{...}{Parameters to ggexp::plot_distributions}
}
\value{
ggplot object
}
\description{
Plot feature distributions from SingleCellExperiment objects
}
\examples{
NULL
}
