% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_heatmap.R
\name{plot_feature_heatmap}
\alias{plot_feature_heatmap}
\title{Plot heatmap of features across cells with annotations}
\usage{
plot_feature_heatmap(
  sce_list,
  features,
  cell_annotations,
  feature_annotations,
  split_cells,
  split_features,
  assay = "logcounts",
  alt_exp = NULL,
  subsample_cells = FALSE,
  show_cell_dend = FALSE,
  show_feature_dend = TRUE,
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{features}{features to use}

\item{cell_annotations}{columns in colData to annotate on heatmap}

\item{feature_annotations}{data frame of annotations related to each feature in `features`}

\item{split_cells}{columns in colData to split columns (cells) by}

\item{split_features}{data frame of groups to split rows (features) by}

\item{assay}{assay in SingleCellExperiment to pull data from}

\item{alt_exp}{altExp in SingleCellExperiment to pull data from}

\item{subsample_cells}{whether to subsample cells to smallest group (if split_cells specified)}

\item{show_cell_dend}{whether to show dendrogram for cells}

\item{show_feature_dend}{whether to show dendrogram for features}

\item{...}{params passed to `ggexp::plot_heatmap`}
}
\value{

}
\description{
Plot heatmap of features across cells with annotations
}
\examples{
NULL
}
