% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gex_qc.R
\name{plot_gex_bivariate_qc}
\alias{plot_gex_bivariate_qc}
\title{Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters}
\usage{
plot_gex_bivariate_qc(
  sce_list,
  x,
  y,
  color = NULL,
  shape = NULL,
  x_filters = NULL,
  y_filters = NULL,
  x_log = TRUE,
  y_log = TRUE,
  text_size = 3,
  facet_rows = NULL,
  facet_columns = NULL,
  facet_type = "grid",
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment object}

\item{x}{numeric column from colData that is in all objects in sce_list}

\item{y}{numeric column from colData that is in all objects in sce_list}

\item{color}{column from colData that is in all objects in sce_list}

\item{shape}{column from colData that is in all objects in sce_list}

\item{x_filters}{list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{y_filters}{list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{x_log}{whether to use log x-axis}

\item{y_log}{whether to use log y-axis}

\item{text_size}{font size for annotations}

\item{facet_rows}{columns to facet on}

\item{facet_columns}{columns to facet on}

\item{facet_type}{either "wrap" or "grid", same as ggplot}

\item{...}{other params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{

}
\description{
Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters
}
\examples{
NULL
}
