% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_features.R
\name{plot_pairwise_features}
\alias{plot_pairwise_features}
\title{Plot pairwise scatterplot of cell-level data}
\usage{
plot_pairwise_features(
  sce_list,
  x,
  y = x,
  assay = "logcounts",
  alt_exp = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  facet_rows = c(),
  facet_columns = c(),
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects to plot}

\item{x}{numeric features to plot on x axis - can be from reducedDims, colData, or assay data}

\item{y}{numeric features to plot on y axis - can be from reducedDims, colData, or assay data}

\item{assay}{assay to obtain data from (ex: counts, logcounts)}

\item{alt_exp}{alternate experiment to obtain data from}

\item{color}{column from reducedDims, colData, or assay data to color by}

\item{shape}{column from reducedDims, colData, or assay data to shape by}

\item{size}{column from reducedDims, colData, or assay data to size by}

\item{facet_rows}{columns from colData to facet on}

\item{facet_columns}{columns from colData to facet on}

\item{...}{other parameters to be passed to ggexp::plot_pairwise_scatterplot}
}
\value{
ggplot object
}
\description{
Plot pairwise scatterplot of cell-level data
}
\examples{
NULL
}
