% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_scatterplot.R
\name{plot_pairwise_scatterplot}
\alias{plot_pairwise_scatterplot}
\title{Plot pairwise scatterplot of cell-level data}
\usage{
plot_pairwise_scatterplot(
  sce_list,
  assay = "logcounts",
  alt_exp = NULL,
  x,
  y = x,
  color = NULL,
  shape = NULL,
  size = NULL,
  facet_rows = c(),
  facet_columns = c(),
  ...
)
}
\arguments{
\item{sce_list}{List of SingleCellExperiment objects to plot}

\item{assay}{Assay to obtain data from (ex: counts, logcounts)}

\item{alt_exp}{Alternate experiment to obtain data from}

\item{x}{Numeric features to plot on x axis - can be from reducedDims, colData, or assay data}

\item{y}{Numeric features to plot on y axis - can be from reducedDims, colData, or assay data}

\item{color}{Column from reducedDims, colData, or assay data to color by}

\item{shape}{Column from reducedDims, colData, or assay data to shape by}

\item{size}{Column from reducedDims, colData, or assay data to size by}

\item{...}{Other parameters to be passed to ggexp::plot_pairwise_scatterplot}
}
\value{
ggplot object
}
\description{
Plot pairwise scatterplot of cell-level data
}
\examples{
NULL
}
