% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_qc.R
\name{plot_vdj_gex_univariate_qc}
\alias{plot_vdj_gex_univariate_qc}
\title{Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters}
\usage{
plot_vdj_gex_univariate_qc(
  sce_list,
  x_filters = NULL,
  vdj_filters = NULL,
  x,
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{x_filters}{list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{x}{numeric column from colData that is in all objects in sce_list}

\item{...}{other params passed into either facet_wrap or facet_grid, depending on facet_type parameter}

\item{y}{discrete column from colData that is in all objects in sce_list to split histograms by}

\item{color}{column from colData that is in all objects in sce_list}

\item{shape}{column from colData that is in all objects in sce_list}

\item{x_log}{whether to use log x-axis}

\item{text_size}{font size for annotations}

\item{facet_rows}{columns to facet on}

\item{facet_columns}{columns to facet on}

\item{facet_type}{either "wrap" or "grid", same as ggplot}
}
\value{

}
\description{
Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters
}
\examples{
NULL
}
