% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_qc.R
\name{plot_vdj_qc}
\alias{plot_vdj_qc}
\title{Plot number of VDJ chains versus colData variable of interest with annotated filters}
\usage{
plot_vdj_qc(sce, y, vdj_doublet_filter, y_filter, y_log = TRUE, y_label = y)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{y}{Numeric column from colData}

\item{vdj_doublet_filter}{Filter returned from filter_vdj_doublets}

\item{y_filter}{Filter result from scater::isOutlier or from scanalysis::filter_gex_outliers}

\item{y_log}{Boolean to use log y-axis}
}
\value{

}
\description{
Plot number of VDJ chains versus colData variable of interest with annotated filters
}
\examples{
NULL
}
