% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano plot with annotations}
\usage{
plot_volcano(
  data,
  fold_change,
  p_value,
  color = NULL,
  label = "gene",
  annotations = c(),
  annotations_if_threshold = c(),
  n_annotate_top = 10,
  p_value_threshold = 0.05,
  fold_change_threshold = 1.5,
  p_value_rank_threshold = 100,
  alpha = 0.5,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  facet_scales = "free",
  facet_switch = NULL,
  nrow = 3
)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{fold_change}{string scalar indicating column containing fold change data (x-axis)}

\item{p_value}{string scalar indicating column containing p-value data (y-axis is -log10(p_value))}

\item{color}{string scalar indicating column for color}

\item{label}{string scalar indicating column for text annotation}

\item{annotations}{string vector indicating features to be annotated}

\item{annotations_if_threshold}{string vector indicating features to be annotated iff they meet the thresholds}

\item{n_annotate_top}{integer indicating the number of top DE genes to annotate}

\item{p_value_threshold}{float indicating maximum p-value}

\item{fold_change_threshold}{float indicating minimum absolute fold change threshold}

\item{p_value_rank_threshold}{integer indicating maximum p-value rank threshold}

\item{alpha}{numeric scalar for alpha of points}

\item{facet_rows}{string vector indicating columns for faceting by row}

\item{facet_columns}{string vector indicating columns for faceting by column}

\item{facet_type}{string scalar that is either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{facet_scales}{string scalar that is either "fixed", "free_x", "free_y", or "free", same as scales argument in facet calls}

\item{facet_switch}{string scalar that is either NULL, "both", "x", or "y", same as switch argument in facet calls}

\item{nrow}{numeric scalar indicating the number of rows in plot, only applies if facet_type == "wrap"}
}
\value{
ggplot2 object
}
\description{
Plot volcano plot with annotations
}
\examples{
NULL
}
