% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gex_qc.R
\name{prepare_gex_scale_limits}
\alias{prepare_gex_scale_limits}
\title{Prepare the limits of plot scale}
\usage{
prepare_gex_scale_limits(data, var, filter)
}
\arguments{
\item{data}{Result from prepare_gex_data}

\item{var}{Feature from data that is being plotted}

\item{filter}{Filter for a sample from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}
}
\value{

}
\description{
Prepare the limits of plot scale
}
\examples{
NULL
}
