% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{read_10x}
\alias{read_10x}
\title{Read 10X gene expression and VDJ data into a SingleCellExperiment object}
\usage{
read_10x(
  gene_expr_path = getwd(),
  vdj_path = NA,
  top_n_barcodes = NULL,
  nonzero_barcodes = FALSE
)
}
\arguments{
\item{gene_expr_path}{path to gene expression data - this path must contain mtx, features, and barcodes file}

\item{vdj_path}{path to vdj data - this path must contain all_contig_annotations.csv}
}
\value{

}
\description{
VDJ data is stored in colData(sce)$vdj for now. This makes it easy to manipulate using the tidyverse for assignment of clonotypes, etc.
}
\examples{
NULL
}
