% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{read_10x_gene_expr}
\alias{read_10x_gene_expr}
\title{Read gene expression matrix from 10X data}
\usage{
read_10x_gene_expr(path)
}
\arguments{
\item{path}{Path to gene expression data - can be full path to the cellranger count output for the sample, or the output path of cellranger count if sample name is specified}

\item{sample_name}{Sample name to be used in traversing the file directory structure if general output path is specified}
}
\value{
SingleCellExperiment object
}
\description{
Read gene expression matrix from 10X data
}
\examples{
NULL
}
