% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{sce_to_seurat}
\alias{sce_to_seurat}
\title{Convert SingleCellExperiment object to Seurat and retain multi-modal data}
\usage{
sce_to_seurat(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Convert SingleCellExperiment object to Seurat and retain multi-modal data
}
\examples{
NULL
}
\keyword{internal}
