% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top_de_genes.R
\name{select_top_de_genes}
\alias{select_top_de_genes}
\title{Select the top DE genes, ranked on either fold change or -log10(p-value)}
\usage{
select_top_de_genes(
  data,
  n = 10,
  fold_change = "fold_change",
  p_value = "p_value",
  groups = c(),
  rank_by = "fold_change",
  only_pos = FALSE
)
}
\arguments{
\item{data}{differential expression results}

\item{n}{number of top DE genes to select}

\item{fold_change}{column containing fold change data (x-axis)}

\item{p_value}{column containing p-value data (y-axis is -log10(p_value))}

\item{groups}{column containing groups to get top DE genes per group (for example, cluster column)}

\item{rank_by}{whether to rank by fold_change or p_value first}

\item{only_pos}{only select features with positive fold change}
}
\value{

}
\description{
Select the top DE genes, ranked on either fold change or -log10(p-value)
}
\examples{
NULL
}
