% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{seurat_to_sce}
\alias{seurat_to_sce}
\title{Convert Seurat object to SingleCellExperiment and retain multi-modal data}
\usage{
seurat_to_sce(seurat)
}
\arguments{
\item{seurat}{Seurat object}
}
\value{

}
\description{
Convert Seurat object to SingleCellExperiment and retain multi-modal data
}
\examples{
NULL
}
