% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_vdj.R
\name{unnest_vdj}
\alias{unnest_vdj}
\title{Get VDJ data with cell-level metadata}
\usage{
unnest_vdj(sce, include_all = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object with a column named vdj containing contig_annotations in colData}
}
\value{
DataFrame with row for each vdj sequence containing cell-level metadata
}
\description{
Get VDJ data with cell-level metadata
}
\examples{
NULL
}
