% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gex_qc.R
\name{annotate_n_genes_expr}
\alias{annotate_n_genes_expr}
\title{Add number of genes expressed per barcode to colData(sce)$n_genes_expr}
\usage{
annotate_n_genes_expr(sce)
}
\arguments{
\item{sce}{}
}
\value{

}
\description{
Add number of genes expressed per barcode to colData(sce)$n_genes_expr
}
\examples{
NULL
}
