% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_identity_frequency.R
\name{convert_identity_frequency_matrix_to_long}
\alias{convert_identity_frequency_matrix_to_long}
\title{Convert an identity frequency matrix}
\usage{
convert_identity_frequency_matrix_to_long(
  data,
  attributes,
  group_by,
  normalize
)
}
\arguments{
\item{data}{data frame to encode}

\item{attributes}{columns to use to define features}

\item{group_by}{columns to use to define groups}

\item{normalize}{normalization method}
}
\value{

}
\description{
Convert an identity frequency matrix
}
\examples{
NULL
}
