% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{.read_10x_vdj}
\alias{.read_10x_vdj}
\title{Read VDJ information from 10X data}
\usage{
.read_10x_vdj(path)
}
\arguments{
\item{path}{Path to vdj data - can be full path to the cellranger vdj output for the sample, or the output path of cellranger vdj if sample name is specified}
}
\value{
DataFrame with vdj information
}
\description{
Read VDJ information from 10X data
}
\examples{
NULL
}
\keyword{internal}
