% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assay_data.R
\name{get_assay_data}
\alias{get_assay_data}
\title{Get assay data from either the main experiment or altExps}
\usage{
get_assay_data(sce, assay, alt_exp = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay}{assay to get data from (counts, logcounts, etc.)}

\item{alt_exp}{alternate experiment to get assay data from}
}
\value{

}
\description{
Get assay data from either the main experiment or altExps
}
\examples{
NULL
}
