% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_qc.R
\name{plot_barcode_qc}
\alias{plot_barcode_qc}
\title{QC plot of UMI rank vs total number of UMIs per barcode}
\usage{
plot_barcode_qc(
  sce_list,
  ambient_rna_filters,
  color = "empty_drops",
  shape = NULL,
  facet_rows = NULL,
  facet_columns = NULL,
  facet_type = "grid",
  facet_scales = "fixed",
  facet_switch = NULL,
  nrow = 2,
  text_size = 1
)
}
\arguments{
\item{sce_list}{Named list of SingleCellExperiment objects}

\item{ambient_rna_filters}{Named list of results from filter_ambient_rna function, one for each SCE object in sce_list - must be the same names and order}

\item{color}{Column from colData that is in all objects in sce_list}

\item{shape}{Column from colData that is in all objects in sce_list}

\item{facet_rows}{Columns to facet on}

\item{facet_columns}{Columns to facet on}

\item{facet_type}{Either "wrap" or "grid", same as ggplot}

\item{text_size}{Font size for annotations}

\item{...}{params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{
ggplot with plot of UMI rank vs total number of UMIs per barcode
}
\description{
QC plot of UMI rank vs total number of UMIs per barcode
}
\examples{
NULL
}
