% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gex_qc.R
\name{plot_qc_distribution}
\alias{plot_qc_distribution}
\title{Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters}
\usage{
plot_qc_distribution(
  sce_list,
  x,
  y = NULL,
  color = NULL,
  shape = NULL,
  x_filters = NULL,
  x_log = TRUE,
  text_size = 3,
  facet_rows = NULL,
  facet_columns = NULL,
  facet_type = "wrap",
  ...
)
}
\arguments{
\item{sce_list}{List of SingleCellExperiment object}

\item{x}{Numeric column from colData that is in all objects in sce_list}

\item{y}{Discrete column from colData that is in all objects in sce_list to split histograms by}

\item{color}{Column from colData that is in all objects in sce_list}

\item{shape}{Column from colData that is in all objects in sce_list}

\item{x_filters}{List of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{x_log}{Boolean to use log x-axis}

\item{text_size}{Font size for annotations}

\item{facet_rows}{Columns to facet on}

\item{facet_columns}{Columns to facet on}

\item{facet_type}{Either "wrap" or "grid", same as ggplot}

\item{...}{params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{

}
\description{
Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters
}
\examples{
NULL
}
