% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduced_dimensions.R
\name{plot_reduced_dimensions}
\alias{plot_reduced_dimensions}
\title{Plot reduced dimensional plot with multiple features}
\usage{
plot_reduced_dimensions(
  sce_list,
  type = "PCA",
  assay = "logcounts",
  alt_exp = NULL,
  features,
  label = NULL,
  alpha = 0.3,
  point_size = 1,
  text_size = 1,
  facet_rows = c(),
  facet_columns = c(".feature"),
  facet_type = "grid",
  ...
)
}
\arguments{
\item{sce_list}{List of SingleCellExperiment objects to plot}

\item{type}{Name of reducedDim attribute to plot}

\item{assay}{Assay to obtain data from (ex: counts, logcounts)}

\item{alt_exp}{Alternate experiment to obtain data from}

\item{features}{Features to plot - can be from reducedDims, colData, or assay data, but note that all must be either numeric or categorical for one plot}

\item{alpha}{Alpha for points}

\item{point_size}{Size of points}

\item{text_size}{Size of font for text annotation}

\item{facet_rows}{Variables from colData to facet on, can also include ".sample" or ".feature" as described below}

\item{facet_columns}{Variables from colData to facet on, can also include ".sample" or ".feature" as described below}

\item{facet_type}{Either "wrap" or "grid", same as ggplot}

\item{...}{Other parameters to be passed to ggexp::plot_pairwise_scatterplot}

\item{label_value}{Boolean to annotate text label for the value - only works if all features are discrete}
}
\value{
ggplot object
}
\description{
Features are min-max normalized per feature, and the range of each feature is annotated per facet to consolidate multiple features into one color scale.
}
\details{
If multiple SingleCellExperiments are provided in the sce_list, and you want to facet by this, you can add ".sample" to one of the faceting variables, as this is implicitly added in.

In almost all cases, you would want to facet by feature. This is by default included in facet_columns, but if you add additional faceting variables to facet_columns, be sure to also include ".feature". To be clear, ".feature" can also be included in facet_rows. This is not included within the code to allow for flexibility if you want features to be faceted in the rows or columns.
}
\examples{
library(scanalysis)
library(scater)

sce = mockSCE() \%>\%
    logNormCounts() \%>\%
    scater::runPCA()

plot_reduced_dimensions(sce_list = list(sample_1 = sce, sample_2 = sce),
                        features = c("Gene_0001", "Gene_0002", "Gene_0003"),
                        facet_columns = ".sample",
                        facet_rows = ".feature",
                        switch = "y")
}
