% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vdj_qc.R
\name{plot_vdj_gex_univariate_qc}
\alias{plot_vdj_gex_univariate_qc}
\title{Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters}
\usage{
plot_vdj_gex_univariate_qc(
  sce_list,
  x_filters = NULL,
  vdj_filters = NULL,
  x,
  facet_rows = NULL,
  facet_columns = NULL,
  ...
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{x_filters}{list of filters for each SCE object in sce_list from scater::isOutlier, or a vector with attribute named thresholds that is a vector with min and max allowed values}

\item{x}{numeric column from colData that is in all objects in sce_list}

\item{facet_rows}{columns to facet on}

\item{facet_columns}{columns to facet on}

\item{...}{other params passed into plot_gex_univariate_qc}
}
\value{

}
\description{
Scatterplot of two features of interest from colData with annotated thresholds and counts based on filters
}
\examples{
NULL
}
