% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano plot with annotations}
\usage{
plot_volcano(
  data,
  fold_change,
  p_value,
  color = NULL,
  label = "gene",
  annotations = c(),
  annotations_if_threshold = c(),
  n_annotate_top = 10,
  p_value_threshold = 0.05,
  fold_change_threshold = 1.5,
  p_value_rank_threshold = 100,
  alpha = 0.5,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  ...
)
}
\arguments{
\item{data}{differential expression results}

\item{fold_change}{column containing fold change data (x-axis)}

\item{p_value}{column containing p-value data (y-axis is -log10(p_value))}

\item{color}{column for color}

\item{label}{column for text annotation}

\item{annotations}{features to be annotated}

\item{annotations_if_threshold}{features to be annotated iff they meet the thresholds}

\item{n_annotate_top}{number of top DE genes to annotate}

\item{p_value_threshold}{maximum p-value}

\item{fold_change_threshold}{minimum absolute fold change threshold}

\item{p_value_rank_threshold}{maximum p-value rank threshold}

\item{alpha}{alpha of points}

\item{facet_rows}{columns for faceting by row}

\item{facet_columns}{columns for faceting by column}

\item{facet_type}{either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{...}{other params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{
ggplot2 object
}
\description{
Plot volcano plot with annotations
}
\examples{
NULL
}
