% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_10x.R
\name{read_10x}
\alias{read_10x}
\title{Read 10X gene expression and VDJ data into a SingleCellExperiment object}
\usage{
read_10x(gene_expr_path = getwd(), vdj_path = NA)
}
\arguments{
\item{gene_expr_path}{Path to gene expression data - can be full path to the cellranger count output for the sample, or the output path of cellranger count if sample name is specified}

\item{vdj_path}{Path to vdj data - can be full path to the cellranger vdj output for the sample, or the output path of cellranger vdj if sample name is specified}

\item{gene_expr_sample_name}{Sample name to be used in traversing the file directory structure if general output path is specified}

\item{vdj_sample_name}{Sample name to be used in traversing the file directory structure if general output path is specified}
}
\value{

}
\description{
VDJ data is stored in colData(sce)$vdj for now. This makes it easy to manipulate using the tidyverse for assignment of clonotypes, etc.
}
\examples{
NULL
}
